chore(refactor): [BACK-1330] Refactor manual scheduling workflow to fix a bug
- Manually added and scheduled items have had its own separate workflow
set up that has a lot in common with the workflow for prospects on the Prospecting
page. Uniting these two workflows into one fixes a UI bug with scheduling -
stories added manually and scheduling straight away do not appear on the sidebar
if they've been scheduled for the next couple of days.

This initial commit refactors all of the business logic out of the `AddProspectForm`
component into a new `AddProspectFormConnector` component, leaving the form component
to be in charge of the form layout only.
