+++ src/curated-corpus/components/AddProspectForm/AddProspectForm.tsx
@@ -1,6 +1,5 @@
-import React, { useState } from 'react';
+import React from 'react';
 import { Box, Grid, LinearProgress } from '@material-ui/core';
-import { DateTime } from 'luxon';
 
 import { FormikHelpers, FormikValues, useFormik } from 'formik';
 import {
@@ -12,31 +11,11 @@ import {
 import { validationSchema } from './AddProspectForm.validation';
 import { useStyles } from './AddProspectForm.styles';
 
-import {
-  ApprovedCuratedCorpusItem,
-  CreateApprovedCuratedCorpusItemInput,
-  CreateScheduledCuratedCorpusItemInput,
-  useCreateApprovedCuratedCorpusItemMutation,
-  useCreateScheduledCuratedCorpusItemMutation,
-  useGetApprovedItemByUrlLazyQuery,
-  useGetUrlMetadataLazyQuery,
-  useUploadApprovedCuratedCorpusItemImageMutation,
-} from '../../../api/generatedTypes';
-import {
-  useNotifications,
-  useRunMutation,
-  useToggle,
-} from '../../../_shared/hooks';
-import { ApprovedItemModal, ScheduleItemModal } from '..';
-import {
-  downloadAndUploadApprovedItemImageToS3,
-  transformFormInputToCreateApprovedItemInput,
-  transformUrlMetaDataToApprovedItem,
-} from '../../helpers/helperFunctions';
-
 interface AddProspectFormProps {
-  onCancel: VoidFunction;
-  toggleAddProspectModal: VoidFunction;
+  onSubmit: (
+    values: FormikValues,
+    formikHelpers: FormikHelpers<any>
+  ) => void | Promise<any>;
 }
 
 /**
@@ -49,157 +28,7 @@ export const AddProspectForm: React.FC<
   const classes = useStyles();
 
   // de-structure props
-  const { onCancel, toggleAddProspectModal } = props;
-
-  // state variable to store the itemUrl field from the form
-  const [itemUrl, setItemUrl] = useState<string>('');
-
-  // state variable to store the approved item built after
-  // getting the item metadata
-  const [approvedItem, setApprovedItem] = useState<ApprovedCuratedCorpusItem>();
-
-  // set up some hooks
-  const { showNotification } = useNotifications();
-
-  // Keep track of whether the "Approve Item" modal is open or not
-  const [approvedItemModalOpen, toggleApprovedItemModal] = useToggle(false);
-  //Keep track of whether the "Schedule this item" modal is open or not.
-  const [scheduleModalOpen, toggleScheduleModal] = useToggle(false);
-
-  // function to toggle both modals
-  const toggleApprovedItemAndProspectModal = () => {
-    toggleApprovedItemModal();
-    toggleAddProspectModal();
-  };
-
-  // prepare the runMutation helper hook
-  const { runMutation } = useRunMutation();
-
-  // prepare mutation hook to upload image to s3
-  const [uploadApprovedItemImage] =
-    useUploadApprovedCuratedCorpusItemImageMutation();
-
-  // prepare mutation hook to create approved item
-  const [createApprovedItemMutation] =
-    useCreateApprovedCuratedCorpusItemMutation();
-
-  // prepare mutation hook to schedule the approved item
-  const [scheduleCuratedItem] = useCreateScheduledCuratedCorpusItemMutation();
-
-  // lazy query to check if url already exists in the corpus
-  const [getApprovedItemByUrl] = useGetApprovedItemByUrlLazyQuery({
-    notifyOnNetworkStatusChange: true,
-    fetchPolicy: 'no-cache',
-    onCompleted: (data) => {
-      const approvedItem = data?.getApprovedCuratedCorpusItemByUrl;
-
-      // show error toast if the url exists already
-      if (approvedItem) {
-        showNotification('This URL already exists in the Corpus', 'error');
-        return;
-      }
-
-      // fetch metadata for url if it doesn't exist in the corpus
-      if (approvedItem === null) {
-        getUrlMetadata({
-          variables: {
-            url: itemUrl,
-          },
-        });
-      }
-    },
-  });
-
-  // lazy query to get metadata for an url
-  const [getUrlMetadata] = useGetUrlMetadataLazyQuery({
-    notifyOnNetworkStatusChange: true,
-    fetchPolicy: 'no-cache',
-    onCompleted: (data) => {
-      // create an approved item object from the url metadata to be consumed by the edit form
-      const approvedItem = transformUrlMetaDataToApprovedItem(
-        data.getUrlMetadata,
-        false
-      );
-
-      // set state variable so that it can be used by the edit form
-      setApprovedItem(approvedItem);
-      // show edit form
-      toggleApprovedItemModal();
-    },
-  });
-
-  // callback which will be called after clicking save on the edit form
-  // it uploads the item's image to S3 first and then calls the mutation to
-  // create a new approved item with s3 image url
-  const createApprovedItem = async (
-    values: FormikValues,
-    formikHelpers: FormikHelpers<any>
-  ): Promise<void> => {
-    // upload item image to s3
-    const s3ImageUrl = await downloadAndUploadApprovedItemImageToS3(
-      values.imageUrl,
-      uploadApprovedItemImage
-    );
-
-    // build an approved item using the helper and replace the imageUrl with the s3ImageUrl
-    const createApprovedItemInput: CreateApprovedCuratedCorpusItemInput = {
-      ...transformFormInputToCreateApprovedItemInput(values),
-      imageUrl: s3ImageUrl,
-    };
-    // call the create approved item mutation
-    runMutation(
-      createApprovedItemMutation,
-      { variables: { data: { ...createApprovedItemInput } } },
-      'Item successfully added to the curated corpus.',
-      (data) => {
-        // set the state variable approvedItem to the newly created approved item
-        // that will be used by the schedule modal
-        data.createApprovedCuratedCorpusItem &&
-          setApprovedItem({
-            ...data.createApprovedCuratedCorpusItem,
-          });
-
-        //close approved item modal
-        toggleApprovedItemModal();
-        // open schedule modal
-        toggleScheduleModal();
-        formikHelpers.setSubmitting(false);
-      }
-    );
-  };
-
-  // callback which will be passed to the ScheduleItem modal and will execute
-  // the mutation to schedule an approved item
-  const onScheduleSave = (
-    values: FormikValues,
-    formikHelpers: FormikHelpers<any>
-  ): void => {
-    // Set out all the variables we need to pass to the mutation
-    const variables: CreateScheduledCuratedCorpusItemInput = {
-      approvedItemExternalId: approvedItem?.externalId!,
-      scheduledSurfaceGuid: values.scheduledSurfaceGuid,
-      scheduledDate: values.scheduledDate.toISODate(),
-    };
-
-    // Run the mutation
-    runMutation(
-      scheduleCuratedItem,
-      { variables },
-      `Item scheduled successfully for ${values.scheduledDate.toLocaleString(
-        DateTime.DATE_FULL
-      )}`,
-      () => {
-        // close schedule modal
-        toggleScheduleModal();
-        // close add prospect modal
-        toggleAddProspectModal();
-        formikHelpers.setSubmitting(false);
-      },
-      () => {
-        formikHelpers.setSubmitting(false);
-      }
-    );
-  };
+  const { onCancel, onSubmit } = props;
 
   // set up formik object for this form
   const formik = useFormik({
@@ -209,16 +38,7 @@ export const AddProspectForm: React.FC<
     validateOnBlur: false,
     validateOnChange: false,
     validationSchema,
-    onSubmit: (values, formikHelpers) => {
-      setItemUrl(values.itemUrl);
-      getApprovedItemByUrl({
-        variables: {
-          url: values.itemUrl,
-        },
-      });
-
-      formikHelpers.resetForm();
-    },
+    onSubmit,
   });
 
   return (
@@ -249,29 +69,6 @@ export const AddProspectForm: React.FC<
             <LinearProgress />
           </Box>
         </Grid>
-      )}
-
-      {approvedItem && (
-        <>
-          <ApprovedItemModal
-            heading="Review Item"
-            isRecommendation={true}
-            isOpen={approvedItemModalOpen}
-            onSave={createApprovedItem}
-            toggleModal={toggleApprovedItemAndProspectModal}
-            approvedItem={approvedItem}
-          />
-          <ScheduleItemModal
-            headingCopy="Optional: schedule this item"
-            approvedItem={approvedItem}
-            isOpen={scheduleModalOpen}
-            toggleModal={() => {
-              toggleScheduleModal();
-              toggleAddProspectModal();
-            }}
-            onSave={onScheduleSave}
-          />
-        </>
       )}
     </>
   );
